用户端功能全部实现，不包括微信支付功能。这个功能因为个人没有资格开通，所以也不打算写，直接修改数据库就ok

+++ src/main/java/com/example/sky/pojo/entity/AddressBook.java
@@ -21,5 +21,5 @@ public class AddressBook {
     private String districtName;
     private String detail;
     private String label;
-    private Short isDefault;
+    private Integer isDefault;
 }
